Update README and PEP8 fixes
