+++ code/simulated_annealing/simple_simulated_annealing.py
@@ -20,34 +20,35 @@ OUTPUT_DIR = os.path.join(
 )
 CFUNCS = ['simp', 'exp']
 
+
 def cli_parser():
     parser = argparse.ArgumentParser(
-        allow_abbrev=False, 
+        allow_abbrev=False,
         formatter_class=argparse.ArgumentDefaultsHelpFormatter
     )
 
     parser.add_argument(
-        '--tcn', action='store', dest='tc_number', type=int, 
+        '--tcn', action='store', dest='tc_number', type=int,
         default=0, help='Test case number'
     )
     parser.add_argument(
-        '--T', action='store', dest='T', type=float, 
+        '--T', action='store', dest='T', type=float,
         default=40, help='Inital temperature'
     )
     parser.add_argument(
-        '--alpha', action='store', dest='alpha', type=float, 
+        '--alpha', action='store', dest='alpha', type=float,
         default=0.99, help='Cooling factor'
     )
     parser.add_argument(
-        '--epoch', action='store', dest='epochs', type=int, 
+        '--epoch', action='store', dest='epochs', type=int,
         default=1000, help='Number of epochs'
     )
     parser.add_argument(
-        '--n-epoch', action='store', dest='N_per_epochs', type=int, 
+        '--n-epoch', action='store', dest='N_per_epochs', type=int,
         default=100, help='Number of iterations per epoch'
     )
     parser.add_argument(
-        '--s', action='store_true', dest='SAVE', 
+        '--s', action='store_true', dest='SAVE',
         help='If true, stores any generated plots and the summary data'
     )
     parser.add_argument(
@@ -56,13 +57,14 @@ def cli_parser():
     )
     parser.add_argument(
         '--ext', action='store', dest='ext', type=str,
-        default='', 
-        help='Add a prefix to the plots, summary_data and summary_log ' \
-            'before saving it'
+        default='',
+        help='Add a prefix to the plots, summary_data and summary_log '
+        'before saving it'
     )
 
     args = parser.parse_args()
     return args
+
 
 def function_calls(f):
     '''
@@ -308,7 +310,7 @@ class SimpleSimulatedAnnealing:
     '''
 
     def __init__(
-        self, func0, x0, T0, alpha, epochs, N_per_epochs, 
+        self, func0, x0, T0, alpha, epochs, N_per_epochs,
         cooling_func='simp', ext='', **kwargs
     ):
         '''
@@ -330,7 +332,7 @@ class SimpleSimulatedAnnealing:
                 Type of cooling function to be used
                 Choose from: ['simp', 'exp']
             ext: (string), default=''
-                Add a prefix to the plots, summary_data and summary_log before 
+                Add a prefix to the plots, summary_data and summary_log before
                 saving it
             **kwargs:
                 Additional arguments for `func0`
@@ -560,12 +562,12 @@ class SimpleSimulatedAnnealing:
                     fname, rt=rt, func0_calls=func0_calls, x0_len=x0_len,
                     permut=permut, x0=x0, cost0=cost0, xf=xf, costf=costf,
                     reduction_in_cost=reduction_in_cost, T0=T0, alpha=alpha,
-                    epochs=epochs, N_per_epochs=N_per_epochs, 
+                    epochs=epochs, N_per_epochs=N_per_epochs,
                     cooling_func=cooling_func
                 )
                 print(f'\nSummary data saved at: {fname}')
         finally:
-            outputFile.close() 
+            outputFile.close()
 
     @staticmethod
     def factorial_apprx(N):
@@ -605,7 +607,7 @@ if __name__ == '__main__':
     args = cli_parser()
 
     # Set-up parameters for the Simulated Annealing Algorithm
-    T0, alpha = args.T, args.alpha 
+    T0, alpha = args.T, args.alpha
     epochs, N_per_epochs = args.epochs, args.N_per_epochs
     SAVE, ext = args.SAVE, args.ext
     cfunc = args.cfunc
@@ -624,7 +626,6 @@ if __name__ == '__main__':
         cooling_func=cfunc, ext=ext
     )
 
-    
     optim_solution.solver_summary(tc_name=tc_name, save=SAVE)
     optim_solution.plot_cost_hist(save=SAVE)
